+++ test/test_plugin.py
@@ -143,6 +143,6 @@ def test_plugin_compute_request_all_optionals(
 
     assert compute_resources.artifact_errors == {'Detour Factors': ''}
 
-    assert len(computed_artifacts) == 25
+    assert len(computed_artifacts) == 26
     for artifact in computed_artifacts:
         assert isinstance(artifact, Artifact)